Add flavor detail (#23)
* Products tests (#20)

* add ProductControllerIntegrationTest

* clean test code

* Add price to details (#21)

* add price and production cost fields

* clean code

* update postman schema

* create and add queries.md

* add indexes to entities

* clean code

* update db docs and queries.md

* update postman json, add orderDate, add create orders batch

* update DataInitializer

* Add product options (#22)

* add flavors entities

* add product flavors with data initializer

* Update data, add ProductExtra mapper, add ex

* rename mappers, add flavors to products

* add flavor details, remove compose key

* add flavor details, remove a compose key

* add flavor selection, add patterns, clean imports

* add flavor selection, add patterns, clean imports

+++ src/main/java/com/mdmc/posofmyheart/domain/models/ProductExtra.java
@@ -3,7 +3,7 @@ package com.mdmc.posofmyheart.domain.models;
 import java.math.BigDecimal;
 
 public record ProductExtra(
-        Integer idExtra,
+        Long idExtra,
         String name,
         BigDecimal price
 ) {
